+++ local/features/src/mlink-video-controller/module-handlers/module-registry.ts
@@ -8,7 +8,6 @@ import { watchHarajukuModuleConfig } from "@/mlink-video-controller/modules/watc
 import { watchMylistSelectorModuleConfig } from "@/mlink-video-controller/modules/watch-mylist-selector-module";
 import { watchTabSessionsModuleConfig } from "@/mlink-video-controller/modules/watch-tab-sessions-module";
 import { thumbnailsFilterModuleConfig } from "@/mlink-video-controller/modules/thumbnails-filter-module";
-import { deletedVideoDetectorModuleConfig } from "@/mlink-video-controller/modules/deleted-video-detector-module";
 
 /**
  * 利用可能なモジュールの登録・管理を行うクラス
@@ -42,7 +41,6 @@ export class ModuleRegistry {
       watchMylistSelectorModuleConfig,
       watchTabSessionsModuleConfig,
       thumbnailsFilterModuleConfig,
-      deletedVideoDetectorModuleConfig,
     ].forEach((config) => {
       this.registerModule(config);
     });